pyproject.toml
[project]
name = "donglora-ai-bot"
version = "0.1.0"
description = "MeshCore AI chatbot — default mesh name 'Orac'"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Stephen Waits" }]
dependencies = [
    "donglora",
    "anthropic",
    "cobs",
    "pyserial",
    "pycryptodome",
    "pynacl",
]

[dependency-groups]
dev = ["ruff", "pytest", "pyright"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.pyright]
include = ["orac", "tests", "ai_bot.py"]
pythonVersion = "3.10"
typeCheckingMode = "standard"
reportMissingTypeStubs = false

[tool.uv]
package = false

[tool.uv.sources]
donglora = { path = "../client-py", editable = true }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "ARG", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["donglora", "orac"]
